add fish support to tmux

+++ tmux/setup_tmux.sh
@@ -2,5 +2,8 @@ mkdir -p ~/.tmux/plugins
 
 git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
 
-rm ~/.tmux.conf
+if [ -e ~/.tmux.conf ]
+then
+    rm ~/.tmux.conf
+fi
 ln -s "${HOME}/dotfiles/tmux/tmux.conf" ~/.tmux.conf
